requirements.txt
asgiref==3.4.1
colorama==0.4.4
Django==3.2.9
Pillow==8.4.0
pytz==2021.3
qrcode==7.3.1
sqlparse==0.4.2



